backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-whois
dnspython
beautifulsoup4
builtwith
tldextract
python-dateutil
python-dotenv